requirements.txt
sounddevice==0.4.8
numpy==1.26.4
webrtcvad==2.0.10
aiohttp==3.9.3
requests==2.31.0
rapidfuzz==3.6.1
boto3==1.34.34
paho-mqtt==1.6.1
